requirements.txt
requests
python-dotenv
tk
pillow